fix: correct issues preventing tests from passing
A number of changes were made to ensure tests pass as expected. It was necessary to call the controller and request handler separately as the request wrapper returns a promise and needed async/await. Removed supertest as a dependency and replaced it with axios. This is more consistent with the frontend dependency choices. Although there is a small performance cost making actual requests to a running server instance is more in line with actual application runtime conditions. Added @jest-mock/express as a dependency to simplify mocking of express request, response and next.