fix: fix logout test which was throwing a non-intuitive error
As logout implements req.session.destroy, manually destroying the user and loggedIn params is superfluous. Calling session.save to update the session object conflicts with session.destroy in a race condition causing tests to sometimes log a non-intuitive prisma update error which does not impact on system behaviour. This commit removes the unneeded code, simplifies the codebase and prevents the confusing logged error.